added health

+++ GunMania_Prototype/Assets/Scripts/SL_Script/Player/sl_BulletScript.cs
@@ -5,7 +5,7 @@ using Photon.Pun;
 
 public class sl_BulletScript : MonoBehaviour
 {
-    public int bulletDmg;
+    public float bulletDmg;
    
     void Update()
     {
